Add initial example for programmatic SVG creation
In order to make sure the lesson is reinforced (because it's a lot to
digest!), I think it would be good to have a small project that
highlights an interesting application: dynamic image creation. In our
case, it would be writing a program to generate simple charts using data
from a small JSON object.